chore: migrate BGP multidiut
