+++ tests/snappi_tests/bgp/test_bgp_outbound_downlink_port_flap.py
@@ -2,12 +2,12 @@ import pytest
 import logging
 from tests.common.helpers.assertions import pytest_require, pytest_assert                            # noqa: F401
 from tests.common.fixtures.conn_graph_facts import conn_graph_facts, \
-     fanout_graph_facts_multidut                                                                    # noqa: F401
+     fanout_graph_facts_multidut                                                                   # noqa: F401
 from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port, \
-     snappi_api, multidut_snappi_ports_for_bgp                                                      # noqa: F401
-from tests.snappi_tests.variables import t1_t2_device_hostnames                                     # noqa: F401
-from tests.snappi_tests.multidut.bgp.files.bgp_outbound_helper import (
-     get_hw_platform, run_bgp_outbound_process_restart_test)                                        # noqa: F401
+     snappi_api, multidut_snappi_ports_for_bgp                                                     # noqa: F401
+from tests.snappi_tests.variables import t1_side_interconnected_port, t1_t2_device_hostnames       # noqa: F401
+from tests.snappi_tests.bgp.files.bgp_outbound_helper import (
+     get_hw_platform, run_bgp_outbound_link_flap_test)                                              # noqa: F401
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams                           # noqa: F401
 
 logger = logging.getLogger(__name__)
@@ -37,14 +37,14 @@ ROUTE_RANGES = [{
             }]
 
 
-def test_bgp_outbound_downlink_process_crash(snappi_api,                                     # noqa: F811
-                                             multidut_snappi_ports_for_bgp,                       # noqa: F811
-                                             conn_graph_facts,                             # noqa: F811
-                                             fanout_graph_facts_multidut,                           # noqa: F811
-                                             duthosts,                                  # noqa: F811
-                                             creds):                                # noqa: F811
+def test_bgp_outbound_downlink_port_flap(snappi_api,                                     # noqa: F811
+                                         multidut_snappi_ports_for_bgp,                       # noqa: F811
+                                         conn_graph_facts,                             # noqa: F811
+                                         fanout_graph_facts_multidut,                           # noqa: F811
+                                         duthosts,
+                                         creds):
     """
-    Gets the packet loss duration on killing certain processes in downlink
+    Gets the packet loss duration on flapping the interconnected port between T1 and downlink in T1 side
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
@@ -52,18 +52,16 @@ def test_bgp_outbound_downlink_process_crash(snappi_api,                        
         conn_graph_facts (pytest fixture): connection graph
         fanout_graph_facts_multidut (pytest fixture): fanout graph
         duthosts (pytest fixture): list of DUTs
-        creds (dict): DUT credentials
+        creds (pytest fixture): DUT credentials
     Returns:
         N/A
     """
+
     snappi_extra_params = SnappiTestParams()
     snappi_extra_params.ROUTE_RANGES = ROUTE_RANGES
     snappi_extra_params.iteration = ITERATION
-    snappi_extra_params.test_name = "T2 Downlink Process Crash"
-    snappi_extra_params.multi_dut_params.process_names = {
-                                                            'swss': "/usr/bin/orchagent",
-                                                            'syncd': "/usr/bin/syncd",
-                                                        }
+    snappi_extra_params.test_name = "T1 Interconnectivity flap"
+
     ansible_dut_hostnames = []
     for duthost in duthosts:
         ansible_dut_hostnames.append(duthost.hostname)
@@ -89,9 +87,13 @@ def test_bgp_outbound_downlink_process_crash(snappi_api,                        
         else:
             continue
 
-    snappi_extra_params.multi_dut_params.host_name = t1_t2_device_hostnames[hw_platform][2]
+    snappi_extra_params.multi_dut_params.flap_details = {
+        'device_name': t1_t2_device_hostnames[hw_platform][0],
+        'port_name': t1_side_interconnected_port[hw_platform]
+    }
+
     snappi_extra_params.multi_dut_params.multi_dut_ports = multidut_snappi_ports_for_bgp
     snappi_extra_params.multi_dut_params.hw_platform = hw_platform
-    run_bgp_outbound_process_restart_test(api=snappi_api,
-                                          creds=creds,
-                                          snappi_extra_params=snappi_extra_params)
+    run_bgp_outbound_link_flap_test(api=snappi_api,
+                                    creds=creds,
+                                    snappi_extra_params=snappi_extra_params)
