+++ tests/snappi_tests/bgp/test_bgp_outbound_uplink_multi_po_flap.py
@@ -6,8 +6,8 @@ from tests.common.fixtures.conn_graph_facts import conn_graph_facts, \
 from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port, \
      snappi_api, multidut_snappi_ports_for_bgp                                                       # noqa: F401
 from tests.snappi_tests.variables import t1_t2_device_hostnames                                     # noqa: F401
-from tests.snappi_tests.multidut.bgp.files.bgp_outbound_helper import (
-     get_hw_platform, run_bgp_outbound_ungraceful_restart, run_dut_configuration)                   # noqa: F401
+from tests.snappi_tests.bgp.files.bgp_outbound_helper import (
+     run_bgp_outbound_uplink_blackout_test, run_dut_configuration, get_hw_platform)                 # noqa: F401
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams                           # noqa: F401
 
 logger = logging.getLogger(__name__)
@@ -17,14 +17,24 @@ pytestmark = [pytest.mark.topology('multidut-tgen')]
 ITERATION = 1
 ROUTE_RANGES = [{
                     'IPv4': [
-                        ['100.1.1.1', 24, 15000],
-                        ['200.1.1.1', 24, 15000]
+                        ['100.1.1.1', 24, 500],
+                        ['200.1.1.1', 24, 500]
                     ],
                     'IPv6': [
-                        ['5000::1', 64, 15000],
-                        ['4000::1', 64, 15000]
+                        ['5000::1', 64, 500],
+                        ['4000::1', 64, 500]
                     ],
-                }]
+                },
+                {
+                    'IPv4': [
+                        ['100.1.1.1', 24, 2500],
+                        ['200.1.1.1', 24, 2500]
+                    ],
+                    'IPv6': [
+                        ['5000::1', 64, 2500],
+                        ['4000::1', 64, 2500]
+                    ],
+            }]
 
 
 def test_dut_configuration(multidut_snappi_ports_for_bgp,                  # noqa: F811
@@ -51,14 +61,14 @@ def test_dut_configuration(multidut_snappi_ports_for_bgp,                  # noq
 
     hw_platform = get_hw_platform(ansible_dut_hostnames)
     if hw_platform is None:
-        pytest_require(False, "Unable to get the hardware platform")
+        pytest_require(False, "Unknown HW Platform")
+    logger.info("HW Platform: {}".format(hw_platform))
 
-    logger.info("hw_platform: {}".format(hw_platform))
     for device_hostname in t1_t2_device_hostnames[hw_platform]:
         if device_hostname not in ansible_dut_hostnames:
             logger.info('!!!!! Attention: {} not in : {} derived from ansible dut hostnames'.
                         format(device_hostname, ansible_dut_hostnames))
-            pytest_assert(False, "Mismatch between the dut hostnames in ansible and in variables.py files")
+            pytest_require(False, "Mismatch between the dut hostnames in ansible and in variables.py files")
 
     for duthost in duthosts:
         if t1_t2_device_hostnames[hw_platform][0] in duthost.hostname:
@@ -69,20 +79,20 @@ def test_dut_configuration(multidut_snappi_ports_for_bgp,                  # noq
             snappi_extra_params.multi_dut_params.duthost3 = duthost
         else:
             continue
-    snappi_extra_params.multi_dut_params.hw_platform = hw_platform
+
     snappi_extra_params.multi_dut_params.multi_dut_ports = multidut_snappi_ports_for_bgp
+    snappi_extra_params.multi_dut_params.hw_platform = hw_platform
     run_dut_configuration(snappi_extra_params)
 
 
-def test_bgp_outbound_uplink_ungraceful_restart(snappi_api,                                     # noqa: F811
-                                                multidut_snappi_ports_for_bgp,                  # noqa: F811
-                                                conn_graph_facts,                             # noqa: F811
-                                                fanout_graph_facts_multidut,                  # noqa: F811
-                                                duthosts,
-                                                creds):                                # noqa: F811
-
+def test_bgp_outbound_uplink_complete_blackout(snappi_api,                                     # noqa: F811
+                                               multidut_snappi_ports_for_bgp,                       # noqa: F811
+                                               conn_graph_facts,                             # noqa: F811
+                                               fanout_graph_facts_multidut,                   # noqa: F811
+                                               duthosts,
+                                               creds):
     """
-    Gets the packet loss duration on issuing ungraceful restart in uplink
+    Gets the packet loss duration on flapping all portchannels in uplink side
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
@@ -90,13 +100,15 @@ def test_bgp_outbound_uplink_ungraceful_restart(snappi_api,                     
         conn_graph_facts (pytest fixture): connection graph
         fanout_graph_facts_multidut (pytest fixture): fanout graph
         duthosts (pytest fixture): list of DUTs
+        creds (pytest fixture): DUT credentials
     Returns:
         N/A
     """
     snappi_extra_params = SnappiTestParams()
     snappi_extra_params.ROUTE_RANGES = ROUTE_RANGES
     snappi_extra_params.iteration = ITERATION
-    snappi_extra_params.test_name = "Uplink Ungraceful Re-start"
+    snappi_extra_params.test_name = "T2 Uplink Complete Blackout"
+    snappi_extra_params.multi_dut_params.BLACKOUT_PERCENTAGE = 100
 
     ansible_dut_hostnames = []
     for duthost in duthosts:
@@ -111,7 +123,7 @@ def test_bgp_outbound_uplink_ungraceful_restart(snappi_api,                     
         if device_hostname not in ansible_dut_hostnames:
             logger.info('!!!!! Attention: {} not in : {} derived from ansible dut hostnames'.
                         format(device_hostname, ansible_dut_hostnames))
-            pytest_assert(False, "Mismatch between the dut hostnames in ansible and in variables.py files")
+            pytest_require(False, "Mismatch between the dut hostnames in ansible and in variables.py files")
 
     for duthost in duthosts:
         if t1_t2_device_hostnames[hw_platform][0] in duthost.hostname:
@@ -120,30 +132,24 @@ def test_bgp_outbound_uplink_ungraceful_restart(snappi_api,                     
             snappi_extra_params.multi_dut_params.duthost2 = duthost
         elif t1_t2_device_hostnames[hw_platform][2] in duthost.hostname:
             snappi_extra_params.multi_dut_params.duthost3 = duthost
-        elif t1_t2_device_hostnames[hw_platform][3] in duthost.hostname:
-            snappi_extra_params.multi_dut_params.duthost4 = duthost
         else:
             continue
 
-    snappi_extra_params.device_name = t1_t2_device_hostnames[hw_platform][1]
     snappi_extra_params.multi_dut_params.hw_platform = hw_platform
     snappi_extra_params.multi_dut_params.multi_dut_ports = multidut_snappi_ports_for_bgp
-
-    run_bgp_outbound_ungraceful_restart(api=snappi_api,
-                                        creds=creds,
-                                        is_supervisor=False,
-                                        snappi_extra_params=snappi_extra_params)
+    run_bgp_outbound_uplink_blackout_test(api=snappi_api,
+                                          creds=creds,
+                                          snappi_extra_params=snappi_extra_params)
 
 
-def test_bgp_outbound_downlink_ungraceful_restart(snappi_api,                                     # noqa: F811
-                                                  multidut_snappi_ports_for_bgp,                  # noqa: F811
-                                                  conn_graph_facts,                             # noqa: F811
-                                                  fanout_graph_facts_multidut,                  # noqa: F811
-                                                  duthosts,
-                                                  creds):                                # noqa: F811
-
+def test_bgp_outbound_uplink_partial_blackout(snappi_api,                                     # noqa: F811
+                                              multidut_snappi_ports_for_bgp,                       # noqa: F811
+                                              conn_graph_facts,                             # noqa: F811
+                                              fanout_graph_facts_multidut,                   # noqa: F811
+                                              duthosts,
+                                              creds):
     """
-    Gets the packet loss duration on issuing ungraceful restart in downlink
+    Gets the packet loss duration on flapping 50% of portchannels in uplink side
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
@@ -151,13 +157,15 @@ def test_bgp_outbound_downlink_ungraceful_restart(snappi_api,                   
         conn_graph_facts (pytest fixture): connection graph
         fanout_graph_facts_multidut (pytest fixture): fanout graph
         duthosts (pytest fixture): list of DUTs
+        creds (pytest fixture): DUT credentials
     Returns:
         N/A
     """
     snappi_extra_params = SnappiTestParams()
     snappi_extra_params.ROUTE_RANGES = ROUTE_RANGES
     snappi_extra_params.iteration = ITERATION
-    snappi_extra_params.test_name = "Downlink Ungraceful Re-start"
+    snappi_extra_params.test_name = "T2 Uplink Partial Blackout"
+    snappi_extra_params.multi_dut_params.BLACKOUT_PERCENTAGE = 50
 
     ansible_dut_hostnames = []
     for duthost in duthosts:
@@ -172,7 +180,7 @@ def test_bgp_outbound_downlink_ungraceful_restart(snappi_api,                   
         if device_hostname not in ansible_dut_hostnames:
             logger.info('!!!!! Attention: {} not in : {} derived from ansible dut hostnames'.
                         format(device_hostname, ansible_dut_hostnames))
-            pytest_assert(False, "Mismatch between the dut hostnames in ansible and in variables.py files")
+            pytest_require(False, "Mismatch between the dut hostnames in ansible and in variables.py files")
 
     for duthost in duthosts:
         if t1_t2_device_hostnames[hw_platform][0] in duthost.hostname:
@@ -181,77 +189,11 @@ def test_bgp_outbound_downlink_ungraceful_restart(snappi_api,                   
             snappi_extra_params.multi_dut_params.duthost2 = duthost
         elif t1_t2_device_hostnames[hw_platform][2] in duthost.hostname:
             snappi_extra_params.multi_dut_params.duthost3 = duthost
-        elif t1_t2_device_hostnames[hw_platform][3] in duthost.hostname:
-            snappi_extra_params.multi_dut_params.duthost4 = duthost
         else:
             continue
 
-    snappi_extra_params.device_name = t1_t2_device_hostnames[hw_platform][2]
     snappi_extra_params.multi_dut_params.hw_platform = hw_platform
     snappi_extra_params.multi_dut_params.multi_dut_ports = multidut_snappi_ports_for_bgp
-
-    run_bgp_outbound_ungraceful_restart(api=snappi_api,
-                                        creds=creds,
-                                        is_supervisor=False,
-                                        snappi_extra_params=snappi_extra_params)
-
-
-def test_bgp_outbound_supervisor_ungraceful_restart(snappi_api,                                     # noqa: F811
-                                                    multidut_snappi_ports_for_bgp,                  # noqa: F811
-                                                    conn_graph_facts,                             # noqa: F811
-                                                    fanout_graph_facts_multidut,                  # noqa: F811
-                                                    duthosts,
-                                                    creds):                                # noqa: F811
-
-    """
-    Gets the packet loss duration on issuing ungraceful restart in supervisor
-
-    Args:
-        snappi_api (pytest fixture): SNAPPI session
-        multidut_snappi_ports_for_bgp (pytest fixture):  Port mapping info on multidut testbed
-        conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts_multidut (pytest fixture): fanout graph
-        duthosts (pytest fixture): list of DUTs
-    Returns:
-        N/A
-    """
-    snappi_extra_params = SnappiTestParams()
-    snappi_extra_params.ROUTE_RANGES = ROUTE_RANGES
-    snappi_extra_params.iteration = ITERATION
-    snappi_extra_params.test_name = "Supervisor Ungraceful Re-start"
-
-    ansible_dut_hostnames = []
-    for duthost in duthosts:
-        ansible_dut_hostnames.append(duthost.hostname)
-
-    hw_platform = get_hw_platform(ansible_dut_hostnames)
-    if hw_platform is None:
-        pytest_require(False, "Unknown HW Platform")
-    logger.info("HW Platform: {}".format(hw_platform))
-
-    for device_hostname in t1_t2_device_hostnames[hw_platform]:
-        if device_hostname not in ansible_dut_hostnames:
-            logger.info('!!!!! Attention: {} not in : {} derived from ansible dut hostnames'.
-                        format(device_hostname, ansible_dut_hostnames))
-            pytest_assert(False, "Mismatch between the dut hostnames in ansible and in variables.py files")
-
-    for duthost in duthosts:
-        if t1_t2_device_hostnames[hw_platform][0] in duthost.hostname:
-            snappi_extra_params.multi_dut_params.duthost1 = duthost
-        elif t1_t2_device_hostnames[hw_platform][1] in duthost.hostname:
-            snappi_extra_params.multi_dut_params.duthost2 = duthost
-        elif t1_t2_device_hostnames[hw_platform][2] in duthost.hostname:
-            snappi_extra_params.multi_dut_params.duthost3 = duthost
-        elif t1_t2_device_hostnames[hw_platform][3] in duthost.hostname:
-            snappi_extra_params.multi_dut_params.duthost4 = duthost
-        else:
-            continue
-
-    snappi_extra_params.device_name = t1_t2_device_hostnames[hw_platform][3]
-    snappi_extra_params.multi_dut_params.hw_platform = hw_platform
-    snappi_extra_params.multi_dut_params.multi_dut_ports = multidut_snappi_ports_for_bgp
-
-    run_bgp_outbound_ungraceful_restart(api=snappi_api,
-                                        creds=creds,
-                                        is_supervisor=True,
-                                        snappi_extra_params=snappi_extra_params)
+    run_bgp_outbound_uplink_blackout_test(api=snappi_api,
+                                          creds=creds,
+                                          snappi_extra_params=snappi_extra_params)
